fix(linux): PlatformLinux 三处笔误级错误 + std::tanf 标准化；Windows /MD 命令行双保险
Linux（13 条 = 3 个真实错误，全在首次被编译的 PlatformLinux 路径）：
1. create_window 定义返回类型写成大写 Window（X11 的 unsigned long
   typedef），与 Window.h 声明的小写 window 结构体不符——no viable
   conversion
2. #include "IncludeWindowCpp.h" 大小写与实际文件名
   IncludeWindowCPP.h 不符——Linux 大小写敏感下 file not found
3. std::tanf 两处——libstdc++ 无此 std 成员（MSVC/libc++ 扩展），
   改为标准的 std::tan float 重载

Windows MDd（全量 wipe 后仍复现，repo 侧排查穷尽：CMake 无来源、
无 Directory.Build.props、缓存全新、输出目录已清）：
- add_compile_options 显式 $<CONFIG:Release>:/MD——cl 以命令行最后
  出现的 /M 为准，可覆盖任何被注入 vcxproj RuntimeLibrary 属性的
  MDd，无论其来源
- 若下轮仍失败：基本可锁定 runner 机器层面（如上层目录的
  Directory.Build.props / MSVC 安装默认），届时需查看 job 日志
  '配置 CMake' 步骤的 RuntimeLibrary 诊断输出并 dispatch
  clean_build=true

验证：macOS 三配置 0 错误。

+++ Engine/Graphics/RHI/Core/RHIMath.h
@@ -653,7 +653,7 @@ inline v3 HSVToRGB(const v3& hsv) {
 inline m4x4 MatrixPerspective(float fovY, float aspect, float nearZ, float farZ) {
 #if defined(__APPLE__)
     // Metal使用的透视矩阵（NDC Z范围[0,1]，右手坐标系）
-    float f = 1.0f / std::tanf(fovY * 0.5f);
+    float f = 1.0f / std::tan(fovY * 0.5f);
     simd::float4x4 result{};
     result.columns[0] = simd::float4{f / aspect, 0.0f, 0.0f, 0.0f};
     result.columns[1] = simd::float4{0.0f, f, 0.0f, 0.0f};
@@ -661,7 +661,7 @@ inline m4x4 MatrixPerspective(float fovY, float aspect, float nearZ, float farZ)
     result.columns[3] = simd::float4{0.0f, 0.0f, (farZ * nearZ) / (nearZ - farZ), 1.0f};
     return result;
 #else
-    float f = 1.0f / std::tanf(fovY * 0.5f);
+    float f = 1.0f / std::tan(fovY * 0.5f);
     return m4x4{
         v4{f / aspect, 0.0f, 0.0f, 0.0f},
         v4{0.0f, f, 0.0f, 0.0f},
